+++ DataAccess/Models/OrderItemsModel.cs
@@ -2,9 +2,9 @@
 
 public class OrderItemsModel
 {
-    public int Id { get; set; }
-    public int OrderId { get; set; }
-    public int ArticleId { get; set; }
+    public long Id { get; set; }
+    public long OrderId { get; set; }
+    public long ArticleId { get; set; }
     public int Quantity { get; set; }
     public decimal Price { get; set; }
     public DateTime CreatedAt { get; set; }
